requirements.txt
py~=1.10.0
colorama~=0.4.3
pip~=20.0.2
keyring~=18.0.1
attrs~=21.2.0
toml~=0.10.2
wheel~=0.34.2
pytest~=6.2.4
pluggy~=0.13.1
iniconfig~=1.1.1
pexpect~=4.6.0
numpy~=1.20.3
pyparsing~=2.4.7
setuptools~=45.2.0
Appium-Python-Client==1.1.0
pytest-html==3.1.1
selenium~=3.141.0
pytz~=2019.3
parse~=1.19.0
urllib3~=1.25.8
SecretStorage~=2.3.1
entrypoints~=0.3
cryptography~=2.8
six~=1.14.0
bcrypt~=3.1.7
pytest-xdist==2.2.1
